requirements.txt
asgiref==3.3.1
Django==3.1.8
django-cors-headers==3.11.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
psycopg2-binary
pytz==2021.1
sqlparse==0.4.1
requests==2.26.0
pandas==1.3.5
sqlalchemy==1.4.35